fix(ui-kit): renamed course card components, deleted unused imports, formatted files
