+++ packages/ui-kit/lib/components/course-card/course-card-list.tsx
