+++ packages/ui-kit/lib/components/course-card/course-stats.tsx
