+++ packages/ui-kit/lib/components/course-card/empty-state.tsx
