+++ packages/ui-kit/lib/index.ts
@@ -19,14 +19,14 @@ export {
   type TextAreaInputProps,
 } from './components/text-areaInput';
 export { ThemeProvider, useTheme } from './contexts';
-export { VisitorCourseCard } from './components/coursecard/visitor-coursecard/visitor-coursecard';
-export { CoachCourseCard } from './components/coursecard/coach-coursecard/coach-coursecard';
-export { StudentCourseCard } from './components/coursecard/student-coursecard/student-course-card';
-export { CourseCreatorCard } from './components/coursecard/course-creator-coursecard/course-creator-card';
+export { VisitorCourseCard } from './components/course-card/visitor-course-card/visitor-course-card';
+export { CoachCourseCard } from './components/course-card/coach-course-card/coach-course-card';
+export { StudentCourseCard } from './components/course-card/student-course-card/student-course-card';
+export { CourseCreatorCard } from './components/course-card/course-creator-course-card/course-creator-card';
 export { ProfileTabs, DragAndDrop, Badge };
-export {CoachBanner} from './components/coach-banner';
-export {CourseCard} from './components/coursecard/course-card';
-export { default  as  BuyCoachingSession, type BuyCoachingSessionProps } from './components/buy-coaching-session';
+export { CoachBanner } from './components/coach-banner';
+export { CourseCard } from './components/course-card/course-card';
+export { default as BuyCoachingSession, type BuyCoachingSessionProps } from './components/buy-coaching-session';
 export { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from './components/accordion';
 export * as contexts from './contexts';
-export {Navbar} from './components/navbar';
+export { Navbar } from './components/navbar';
